+++ app/src/main/java/org/simple/clinic/medicalhistory/newentry/NewMedicalHistoryScreen.kt
@@ -3,23 +3,32 @@ package org.simple.clinic.medicalhistory.newentry
 import android.content.Context
 import android.os.Parcelable
 import android.util.AttributeSet
+import android.view.View
 import android.widget.RelativeLayout
 import androidx.appcompat.app.AppCompatActivity
+import androidx.transition.TransitionManager
+import com.google.android.material.transition.MaterialFade
 import com.jakewharton.rxbinding3.view.clicks
 import io.reactivex.Observable
 import io.reactivex.rxkotlin.cast
 import io.reactivex.subjects.PublishSubject
 import io.reactivex.subjects.Subject
+import org.simple.clinic.R
 import org.simple.clinic.ReportAnalyticsEvents
+import org.simple.clinic.appconfig.Country
+import org.simple.clinic.databinding.ListMedicalhistoryHypertensionTreatmentBinding
 import org.simple.clinic.databinding.ScreenNewMedicalHistoryBinding
 import org.simple.clinic.di.injector
 import org.simple.clinic.medicalhistory.Answer
+import org.simple.clinic.medicalhistory.Answer.No
+import org.simple.clinic.medicalhistory.Answer.Yes
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.DIAGNOSED_WITH_DIABETES
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.DIAGNOSED_WITH_HYPERTENSION
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.HAS_HAD_A_HEART_ATTACK
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.HAS_HAD_A_KIDNEY_DISEASE
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.HAS_HAD_A_STROKE
+import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.IS_ON_HYPERTENSION_TREATMENT
 import org.simple.clinic.medicalhistory.SelectDiagnosisErrorDialog
 import org.simple.clinic.mobius.MobiusDelegate
 import org.simple.clinic.mobius.ViewRenderer
@@ -52,7 +61,11 @@ class NewMedicalHistoryScreen(
   @Inject
   lateinit var effectHandlerFactory: NewMedicalHistoryEffectHandler.Factory
 
+  @Inject
+  lateinit var country: Country
+
   private var binding: ScreenNewMedicalHistoryBinding? = null
+  private var hypertensionTreatmentBinding: ListMedicalhistoryHypertensionTreatmentBinding? = null
 
   private val toolbar
     get() = binding!!.toolbar
@@ -81,6 +94,21 @@ class NewMedicalHistoryScreen(
   private val hypertensionDiagnosisView
     get() = binding!!.hypertensionDiagnosisView
 
+  private val hypertensionTreatmentContainer
+    get() = binding!!.hypertensionTreatmentContainer
+
+  private val scrollView
+    get() = binding!!.scrollView
+
+  private val hypertensionTreatmentChipGroup
+    get() = hypertensionTreatmentBinding!!.chipGroup
+
+  private val hypertensionTreatmentYesChip
+    get() = hypertensionTreatmentBinding!!.yesChip
+
+  private val hypertensionTreatmentNoChip
+    get() = hypertensionTreatmentBinding!!.noChip
+
   private val questionViewEvents: Subject<NewMedicalHistoryEvent> = PublishSubject.create()
 
   private val events: Observable<NewMedicalHistoryEvent> by unsafeLazy {
@@ -90,9 +118,9 @@ class NewMedicalHistoryScreen(
         .cast<NewMedicalHistoryEvent>()
   }
 
-  private val uiRenderer: ViewRenderer<NewMedicalHistoryModel> = NewMedicalHistoryUiRenderer(this)
-
   private val mobiusDelegate: MobiusDelegate<NewMedicalHistoryModel, NewMedicalHistoryEvent, NewMedicalHistoryEffect> by unsafeLazy {
+    val uiRenderer: ViewRenderer<NewMedicalHistoryModel> = NewMedicalHistoryUiRenderer(this, country)
+
     MobiusDelegate.forView(
         events = events,
         defaultModel = NewMedicalHistoryModel.default(),
@@ -103,6 +131,13 @@ class NewMedicalHistoryScreen(
     )
   }
 
+  private val hypertensionContainerFade by unsafeLazy {
+    MaterialFade().apply {
+      duration = 150L
+      addTarget(hypertensionTreatmentContainer)
+    }
+  }
+
   override fun onFinishInflate() {
     super.onFinishInflate()
     if (isInEditMode) {
@@ -110,11 +145,21 @@ class NewMedicalHistoryScreen(
     }
 
     binding = ScreenNewMedicalHistoryBinding.bind(this)
+    hypertensionTreatmentBinding = ListMedicalhistoryHypertensionTreatmentBinding.bind(binding!!.hypertensionTreatmentContainer)
 
     context.injector<Injector>().inject(this)
 
     toolbar.setNavigationOnClickListener {
       router.pop()
+    }
+
+    hypertensionTreatmentChipGroup.setOnCheckedChangeListener { _, checkedId ->
+      val answer = when (checkedId) {
+        R.id.yesChip -> Yes
+        R.id.noChip -> No
+        else -> Answer.Unanswered
+      }
+      questionViewEvents.onNext(NewMedicalHistoryAnswerToggled(IS_ON_HYPERTENSION_TREATMENT, answer))
     }
 
     post {
@@ -130,6 +175,7 @@ class NewMedicalHistoryScreen(
   override fun onDetachedFromWindow() {
     mobiusDelegate.stop()
     binding = null
+    hypertensionTreatmentBinding = null
     super.onDetachedFromWindow()
   }
 
@@ -210,6 +256,21 @@ class NewMedicalHistoryScreen(
 
   override fun hideNextButtonProgress() {
     nextButton.setButtonState(Enabled)
+  }
+
+  override fun showHypertensionTreatmentQuestion(answer: Answer) {
+    TransitionManager.beginDelayedTransition(scrollView, hypertensionContainerFade)
+
+    hypertensionTreatmentContainer.visibility = View.VISIBLE
+    hypertensionTreatmentYesChip.isChecked = answer == Yes
+    hypertensionTreatmentNoChip.isChecked = answer == No
+  }
+
+  override fun hideHypertensionTreatmentQuestion() {
+    TransitionManager.beginDelayedTransition(scrollView, hypertensionContainerFade)
+
+    hypertensionTreatmentContainer.visibility = View.GONE
+    hypertensionTreatmentChipGroup.clearCheck()
   }
 
   interface Injector {