+++ app/src/main/java/org/simple/clinic/medicalhistory/newentry/NewMedicalHistoryUiRenderer.kt
@@ -1,5 +1,6 @@
 package org.simple.clinic.medicalhistory.newentry
 
+import org.simple.clinic.appconfig.Country
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.DIAGNOSED_WITH_DIABETES
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.DIAGNOSED_WITH_HYPERTENSION
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.HAS_HAD_A_HEART_ATTACK
@@ -8,7 +9,8 @@ import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.HAS_HAD_A_STROKE
 import org.simple.clinic.mobius.ViewRenderer
 
 class NewMedicalHistoryUiRenderer(
-    private val ui: NewMedicalHistoryUi
+    private val ui: NewMedicalHistoryUi,
+    private val country: Country
 ) : ViewRenderer<NewMedicalHistoryModel> {
 
   override fun render(model: NewMedicalHistoryModel) {
@@ -33,11 +35,21 @@ class NewMedicalHistoryUiRenderer(
           ui.renderAnswerForQuestion(DIAGNOSED_WITH_DIABETES, hasDiabetes)
         }
       }
+
+      renderHypertensionTreatmentQuestion(model)
     }
 
     ui.showDiagnosisRequiredError(model.showDiagnosisRequiredError)
 
     renderNextButton(model)
+  }
+
+  private fun renderHypertensionTreatmentQuestion(model: NewMedicalHistoryModel) {
+    if (country.isoCountryCode == Country.INDIA && model.diagnosedWithHypertension) {
+      ui.showHypertensionTreatmentQuestion(model.ongoingMedicalHistoryEntry.isOnHypertensionTreatment)
+    } else {
+      ui.hideHypertensionTreatmentQuestion()
+    }
   }
 
   private fun renderNextButton(model: NewMedicalHistoryModel) {